app: Make ::prepare-shutdown a GObject signal
Using the Signals module's signal methods with GObject-derived classes
has always been awkward, as both emitter and consumer of signals need to
be aware of the difference and use dedicated methods. In newer classes,
we already make use of Lang.Class' ability to define GObject signals,
which makes the last remaining case of mixing signal systems even more
awkward - time to remove the old hack and make it a GObject signal.