utils: Match any URI scheme
Since commit 651c71fed54, we don't use a fixed list of URI schemes
to linkify, but query the system for the list of supported schemes.
Unfortunately this requires access to the list of all installed
applications, which we don't have when running in a flatpak sandbox.
We could easily fix the 99% case by always adding http(s) to the
list, but as unsupported schemes should be extremely rare in the
wild, we can just live with the handful of false positives and
linkify any URI scheme.

https://bugzilla.gnome.org/show_bug.cgi?id=772655