Set the username to be the same as the nickname
Connecting to ZNC fails with a NetworkError if username is wrong. By
default telepathy-idle uses the system login when there is no username
set on the account, make the assumption that the nickname is more likely
to match what the user wants as username and that most users don't want
to have a separate username from nickname.

https://bugzilla.gnome.org/show_bug.cgi?id=762481