appNotification: Dismiss undo notifications on shutdown
When the user removes an account, we only hide it initially to give
the user the chance to undo the action. Currently when we quit while
the undo notification is still shown, the account gets stuck in the
hidden state - make sure to properly dismiss the notification in that
case so that the hidden account is actually removed.

https://bugzilla.gnome.org/show_bug.cgi?id=763872