joinDialog: Include room list in join dialog
Currently users have to know exactly which rooms they want to join
and type in the exact name (either in the join dialog or as parameter
to the /join command). It is often more convenient to browse the list
of existing rooms and pick from that instead, so include that list
in the join dialog.

https://bugzilla.gnome.org/show_bug.cgi?id=763200