connections: Always hide default port
When filling the details dialog from an existing account, we do not
append the port to the server address if it matches the default port
for unencrypted connections - this makes sense as the information is
redundant and there's a good chance that the user did not actually
specify the port in the first place. However not doing the same for
encrypted connections means that the "secure connection" checkbox
unexpectedly affects the server entry as well: Without a specified
port, unchecking the box will append the default SSL port. Never
appending the respective default port is less likely to result in
confusing behavior, so do that instead.