application: Try alternative servers on connection failures
Predefined networks can have more than a single server, so when a
connection fails, we have a list of servers we can fall back to
instead of failing immediately. This is particularly important as
we hide the network details in the properties dialog for predefined
connections, so users cannot manually change the server themselves.

https://bugzilla.gnome.org/show_bug.cgi?id=762335