networksManager: Prioritize favorite networks
Using the "favorite" property enables us to flag specific
networks as important. We want this property because the
networks list is long and naturally, we want to put the
favorite ones on top of the list. Furthermore, the current
designs for initial setup use a reduced list of networks,
thus prioritization would make it easier to display the new
list.

Moreover, the sorting function is changed in order to account
for the favorite proprty first, and then the alphabetic order.

https://bugzilla.gnome.org/show_bug.cgi?id=762733