chatView: Make auto-scrolling more robust
We currently implement auto-scrolling by keeping track of the
scroll position that corresponds to the bottom, and checking
whether it matches the current scroll position. This used to
work (mostly), until we started to animate auto-scrolling in
commit 65f9c8a07 - now if more text is added to the log while
the animation is still ongoing (another message, or even just
a timestamp), the check will fail and stop auto-scrolling
until the user manually scrolls to the bottom again.
To fix this, simplify the code to keep track of whether we
should auto-scroll instead of a particular scroll position:
We simply disable auto-scrolling when the user scrolls up,
and re-enable it when reaching the bottom edge. This should
be much more robust, and probably fixes tons of other edge
cases in addition to the issue outlined above.

https://bugzilla.gnome.org/show_bug.cgi?id=765129