pasteManager: Use connect_after() to receive drag data
GTK+'s text widgets already have built-in DND support which may conflict
with our own drop target support. Currently that is not an issue, as the
chat log is not editable, which disables the built-in support. However we
want to extend the entry's built-in drop target support as well, where
this will be an issue - in particular as it should use either the regular
or our custom support.
Running our custom GtkWidget::drag-data-received handler after the object
method handler will allow classes to do that by either making the built-in
handler a NOOP or stopping the signal emission.

https://bugzilla.gnome.org/show_bug.cgi?id=766068