chatView: Add shortcuts to jump to top/bottom
GtkTextView's built-in bindings, <ctrl>Home and <ctrl>End, don't work,
as they are used by the first-room/last-room application shortcuts. As
we hide the cursor and don't have horizontal scrolling in the chat log,
the built-in bindings for jumping to the beginning/end of the current
line aren't useful, so use them to move to the top/bottom of the
chatlog instead.

https://bugzilla.gnome.org/show_bug.cgi?id=755549