networksManager: Load initial networks list synchronously
Loading a file from a resource is super-fast as it does not involve IO,
so there's no real reason to do the operation asynchronously, in
particular as we initialize the networks manager singleton before any UI
is shown. It does open the door for race conditions on startup though,
which we can easily avoid by loading the initial list synchronously.

https://bugzilla.gnome.org/show_bug.cgi?id=766935