chatView: Track contacts associated with nicks and base status on them
Until now, the status of a nickname was updated each time a status
change was observed (join, left, rename, etc.). Due to the fact that
the 'basenick' is used to identify a user tag, multiple users sharing
the same 'basenick' also share the same tag, therefore if one of
the users leaves, then both of them would be marked as offline.

To fix, keep track of the actual contacts corresponding to each
'basenick'. Users are marked as online if there is at least one
contact associated with their 'basenick', and are marked offline
otherwise.

https://bugzilla.gnome.org/show_bug.cgi?id=763868