chatView: Mark all users as offline upon connection loss
We currently update tracked contacts when a connection is
established, but we don't update their status when the
connection is lost. The problem with this is that
whenever an unexpected connection failure occurs, the
color of each nick remains the same as before the connection
failure (the color is not updated). Therefore, upon
reconnecting, some users will have the wrong color set (i.e.
if the user left before we reconnected, he/she will still
be marked as online).

To fix this, each time there is no channel set, mark all users
as offline.

https://bugzilla.gnome.org/show_bug.cgi?id=763868