userList: Use GtkScrolledWindow:max-content-height
Since commit 025ec4172166, we don't expand the user list popover
indefinitely, but limit its height to eight members. This is currently
implemented by not expanding the popover and setting a minimum height
on the content.
GTK+ recently introduced a GtkScrolledWindow:max-content-height property
that unfortunately broke our code, as the popover ends up being expanded
again. However using the new :max-content-height fixes the issue and
reflects the intention of the code better anyway, so just do that.