chatroomManager: Don't use tp timestamp for presenting window
Translating the timestamp from a telepathy request to an X11
timestamp is not entirely reliable (not least because we are
faking TP_USER_ACTION_TIME_CURRENT_TIME due to limitations
in JS's handling of large numbers), so there's a fair chance
that a more recent event causes the window manager to deny
our focus request. We can expect GTK+ to do a better job at
keeping track of relevant timestamps (including remote calls
to the org.freedesktop.Application interface), so simply
activate the application to present the window without an
explicit timestamp.