app: Add option to run in background
Like other messaging applications, IRC clients don't require constant
attention from the user, but can sit in the background and notify the
user when an important message was received. We currently don't have
any explicit support for this mode, so users need to ignore the window
the best they can (by moving it to another workspace or keeping it at
the bottom of the window stack). This is obviously not great, but neither
are status icons that have traditionally been used by applications to
offer this mode.

Instead, implement a pattern for background applications that was agreed
upon with the GNOME design team:

When set up to run in the background, the application:
 - sets itself up to start in the background on login
 - keeps running after the last window has been closed
 - can be stopped via the 'quit' action in the application menu

https://bugzilla.gnome.org/show_bug.cgi?id=770750