chatView: Consider opacity style property for theme colors
GTK+ has supported an 'opacity' CSS property for a while, and designers
have used it to dim elements (namely: .dim-label). This is not reflected
by the return value of gtk_style_context_get_color() though, which returns
the exact color as defined in the theme. To correctly get the intended
appearance, we need to apply the opacity to the returned color.

https://bugzilla.gnome.org/show_bug.cgi?id=768381