entryArea: Filter out BACKDROP state from nick button
We want the button to blend in with the chat log, however when the
toplevel window doesn't have focus, active nicks in the log keep
their color, while the nick button (and thus the user's own nick)
is grayed out and dimmed. Work around this issue by filtering out
the BACKDROP state altogether.

https://bugzilla.gnome.org/show_bug.cgi?id=768381