room: Use nick matching to filter out mentions from self
We do not want to highlight mentions in messages we sent ourselves.
To filter those out, we currently compare the message's :sender
property with the connection's :self-contact. This is another case
where our message highlighting is tied to being connected, so in
order to enable us to highlight mentions in logs as well, try to
match the sender's :alias property against the self-nick instead.

https://bugzilla.gnome.org/show_bug.cgi?id=768613