room: Change should_highlight() parameters
The should_highlight() method currently takes a TpMessage parameter,
which we get from the channel when a message is sent or received.
We don't get TpMessages from the logger though, so change method to
take sender and message text separately as strings instead, to make
it usable for log messages as well.

https://bugzilla.gnome.org/show_bug.cgi?id=768613