chatView: Highlight mentions in the backlog
We generally try to minimize differences between messages that we
sent or received, and messages that we fetched from the logs.
In other words, when restarting polari, the state after the restart
should match the one before as closely as possible. Besides status
messages that are not logged, the most glaring difference for now is
that we don't restore message highlights in the backlog. However
highlighting mentions is not only useful in the active session, it
can be just as helpful while skimming through the backlogs, so not
having them in logs is actually worse than introducing some visual
inconsistency. Luckily PolariRoom's should_highlight_message()
is now capable of handling this case, so we can finally fix this.

https://bugzilla.gnome.org/show_bug.cgi?id=768613