room: Set self-nick in the offline case
We keep track of the connection's self-contact property to set
the nick we use to decide whether a particular message should be
highlighted. Obviously this can only work while we are connected,
so if we want to also start highlighting messages that are loaded
while disconnected, we need to use a different property to match
messages against - the account's nickname is probably the best we
have in this case.

https://bugzilla.gnome.org/show_bug.cgi?id=768613