room: Move casefolding into match function
On systems where strcasestr() is not available, we need to casefold
the text we search in for a case insensitive result. As we are about
to re-use the match helper function, it makes sense to move the
casefolding there instead of leaving it to all callers. It also
means casefolding is not longer tied to a TpMessage parameter that
soon will go away.

https://bugzilla.gnome.org/show_bug.cgi?id=768613