chatView: Re-enable auto-scrolling after reading mention at bottom
We stop auto-scrolling when a message is highlighted, to keep it
visible on screen. However once the message has been read (e.g.
by switching to the room in question), it makes sense to turn
auto-scrolling back on if the message is at the end of the buffer,
as the view is already scrolled to the bottom in this case.

https://bugzilla.gnome.org/show_bug.cgi?id=765129