userDetails: add notification label and notification icon in the ui of the UserDetails class. userDetails: labels now have the same font size and icons are dimmed.
userDetails: add the notifications-enabled property and bind it to the notifyIcon and notifyLabel visible property.. Make the widgets invisible from the .ui (will be set to true later). Set notificationLabel text to 'will notify...'.

userDetails: fullnameLabel has ellipsise and max-witdh-chars set. userDetails: set the notificationLabel in the .ui file and make it translatable. userDetails: make labels translatable and remove unneeded comment.

userDetails: fix the fullnameLabel, lastHeader and lastLabel
labels in UserDetails. Font size is normal, without the
<small> markup and there is a single space (horizontally)
between the lastHeader and lastLabel.

userDetails: update the max-width-chars property of the
fullname label in the UserDetails class