mainWindow: Do not call show_all() from _init()
The expected behavior of widgets in GTK+ is to show all internal
children that should be visible when the widget is shown, but
leave its own visibility to the parent (or the code that created
the widgets). Not following that pattern is confusing for any
widget, but much worse for toplevel windows where the widget's
visibility directly corresponds to the visibility on screen.