userTracker: Minor cleanup
A contact is unique within a room, so if we remove a contact from
a room mapping, we can just search by alias and don't need the
room name.
The same doesn't apply when we remove a contact from the global
contacts list, where the same contact can be added for multiple
rooms. But then we only really care about the number of contacts
for a particular baseNick, so it doesn't matter if we just remove
the first match ...