mainWindow: Manage active room via an :active-room property
The ChatroomManager class hasn't aged very well, and has become
increasingly muddy with regard to what it actually does and how
it relates to other components (in particular the application):

 - ChatroomManager manages joined rooms, except updating
   which rooms should be joined is done by the application

 - ChatroomManager handles incoming channel requests from
   mission-control, but our own channel requests are made
   by the application

 - managing the list of rooms means we want it available as
   early as possible to other components; being a Telepathy
   handler and observer means we only want to start it when
   we should establish network connections

We will untangle that mess over the next couple of commits, starting
with something simple: managing the active room. Code-wise it would
be convenient to keep using an always available singleton for this,
with application being the obvious choice. However conceptually, the
window is better suited to manage the active room, as the whole concept
is tied to some visual representation - being "active" really means
being the currently selected room in the window. If we are ever going
to support more than a single main window, we surely don't want all
windows to show the same room. Similarly, if we allow running in the
background without any windows, the idea that any of the joined rooms
was active seems silly. So move tracking of the active room from the
ChatroomManager into the window.

https://bugzilla.gnome.org/show_bug.cgi?id=769655