app: Stop assuming a single main window
We will eventually allow opening multiple windows, to make it possible
to observe different rooms at the same time (e.g. via tiling/multiple
monitors) - obviously the current assumption of a single main window
conflicts with that, so adapt the code to either use the :active-window
or iterate over all windows as appropriate.