app: Split actions between app and window
Currently all our actions are added to the application, though
conceptually some of them are clearly associated with a particular
window. In a single-window application, this doesn't really matter,
and the approach has some advantages, as all actions are already
available when the UI is constructed. However in order to allow
multiple windows, we won't get away with this any longer, so add
window-specific actions to the window.