mainWindow: Initialize sidebar visibility
The code currently assumes that the window is created before accounts are
initialized, which is the case in the common case that the application is
activated immediately. However when started through an action (for example
when clicking a persistent notification), accounts may already be ready
when we call activate() ourselves, and the sidebar remains hidden (until
the user adds another account). Fix this by making sure the visibility
is properly initialized.

https://bugzilla.gnome.org/show_bug.cgi?id=769592