roomList: Connect channel signals immediately if possible
We use the channel's ::message-received and ::pending-message-removed
signals to mark rooms as active/inactive and update the highlight
counter. Currently we are only connecting to those signals when the
room's channel changed, so we fail to update rows for rooms that
already had a channel when added to the list. While this is not the
common case, it can still happen, e.g. when someone initiates a
private chat with us or when re-launched by mission-control after
a crash.

https://bugzilla.gnome.org/show_bug.cgi?id=768909