chatView: Make sure we insert something when fetching backlog
When the user triggers loading of more logs, the expectation is that
indeed more logs will be inserted in the view. However in order to
only insert nicks once for consecutive messages from the same user,
we keep all log events from the same nick as the first one pending
until we fetch the next batch of log events. As a result we end up
not inserting anything when all events have the same sender. To
meet user expectation as well as maintain our grouping, just fetch
more backlogs immediately in case we didn't insert any logs.

https://bugzilla.gnome.org/show_bug.cgi?id=768907