chatView: Split out _createMessage() method
As we display both Tp.Messages we send/receive and Tpl.Events we
get from logs, we use a small intermediate object that abstracts
the differences when passed to insertMessage(). Splitting the
current code to create those objects into a helper methods makes
it easier to keep the two code paths in sync, plus we'll soon
use it from more places.

https://bugzilla.gnome.org/show_bug.cgi?id=768907