chatView: Store pending logs as messages
We are not using anything from log events besides the properties
we have in the data object we pass to insertMessage(), so we can
just create those immediately and store them instead of the log
events themselves. As we are about to insert pending Tp.Messages
together with logs, this means we can operate on a single data
type.

https://bugzilla.gnome.org/show_bug.cgi?id=768907