chatView: Split out helper function for clarity
When we display consecutive messages from the same sender in the
chat log, we only insert the nick for the first one. This is trivial
for regular messages we append to the end, but a bit trickier for
log messages where we move backwards. To get the grouping right,
we don't insert all retrieved logs immediately, but only from the
first message that groups differently than the oldest message we
currently have from logs. This works fine, but having the code
directly in insertPendingLogs() distracts a bit from the main
functionality, plus a separate function can be simplified by
using returns instead of nesting if blocks.

https://bugzilla.gnome.org/show_bug.cgi?id=768907