chatView: Move handling of 'highlight' marks into insertMessage()
We pause autoscrolling while we have unread highlighted messages.
Currently we create the marks we use to implement this for messages
passed to _insertTpMessage(), which is fine as those are the only
ones that can have an unread state. However we'll soon insert
messages that were pending when we joined a room together with
log events (i.e. from _insertPendingLogs()), so moving the handling
of 'highlight' marks into _insertMessage() allows for sharing that
code.

https://bugzilla.gnome.org/show_bug.cgi?id=768907