roomList: Add "Connect" menu item
The existing reconnect action only works for accounts that are not
offline (by user request). However there are cases where initiating
a connection can be useful, mostly when not joining channels through
the UI, but via an external service like webirc or bouncers. To
account for this, show a "Connect" item when the "Reconnect" action
isn't available.

https://bugzilla.gnome.org/show_bug.cgi?id=769192