build: Check for runtime dependencies
So far we only include checks for dependencies needed at compile-time
in configure,  which is inconvenient for users who try to build polari
themselves. autoconf-archive now ships an AX_CHECK_GIRS_GJS macro we can
use to check for girs we depend on at runtime without cluttering compiler
and linker flags, which was the main reason for not checking for runtime
dependencies.

https://bugzilla.gnome.org/show_bug.cgi?id=769280