accountsMonitor: Mirror ::account-enabled/disabled signals
To get notified about accounts being enabled or disabled rather than
added/removed, we currently use the Tp.AccountManager signals directly.
That's not terrible, but having the signals on the AccountsMonitor
instead is a bit more convenient, and means we only get notifications
about accounts we actually care about (that is, no GTalk/Jabber or
other non-IRC ones).

https://bugzilla.gnome.org/show_bug.cgi?id=769582