chatroomManager: Split out room management
Originally rooms were tightly coupled with channels - a room was created
and added when we observed a channel, and removed when the channel was
invalidated - so doing room management in the telepathy observer/handler
made a lot of sense. But nowadays where we show disconnected rooms, the
only point where room management overlaps with channel handling is where
we match a channel to a (usually existing) room. It is time to split out
room management into its own class, including maintaining the list of
saved channels (where updating the list got split from restoring rooms
when the code moved into the application).

https://bugzilla.gnome.org/show_bug.cgi?id=769655