application: Initialize actions as early as possible
Until recently, some of the action create hooks were using some of the
singletons that are initialized on startup, so we had to create those
first before adding the actions. Now that this is no longer the case,
we can reorder the code so that all application actions are always
available, including to the singletons created on startup.

https://bugzilla.gnome.org/show_bug.cgi?id=769655