telepathyClient: Connect accounts explicitly before creating channels
We currently rely on mission-control automatically establishing server
connections when we request a channel. However we will soon support
NickServ identification, which means we may need to send an "identify"
message before joining any channels. This will be easier to achieve if
connecting to the server and joining channels are separate operations,
so stop relying on automatic presence changes and connect accounts
explicitly before creating any channels.
As a side effect, externally configured channels (a.k.a. autojoin) will
now work as expected, as we do not filter out enabled accounts without
rooms.

https://bugzilla.gnome.org/show_bug.cgi?id=751575