telepathyClient: Stop using the singleton pattern
The pattern made a lot of sense while many components needed to access it
for its room management, but now that it has been reduced to its function
as telepathy client, it is cleaner to make it a regular class that is
instantiated by the application when its functionality is actually
needed.
This also allows us to implement Telepathy.BaseClient directly instead of
using a helper class, and to get rid of the awkward lateInit() split.

https://bugzilla.gnome.org/show_bug.cgi?id=769655