app: Fix restored account names
When a connection fails because the used account is already connected,
we retry the connection up to three times appending '_' to the account
name on each attempt, and restore the original nick after either
connecting successfully or giving up. However the nickname property we
currently use to restore the account name is ill-suited for that
purpose, as it doesn't necessarily match the account parameter, but
rather the nickname that was active when last connected (which may
be a temporary nick picked by the user like 'nick-afk', or indeed
have the underscore suffix we want to revert).
We want to restore the original account parameter, so actually use that
parameter rather than a property that may or may not match ...

https://bugzilla.gnome.org/show_bug.cgi?id=769583