telepathyClient: Disconnect accounts when network is unavailable
We currently only request a connection when network is available, to
avoid showing busy spinners for pointless connection attempts. This
works when we start offline and the network becomes available, however
there's no point either in trying to re-establish a connection when
the network goes away, so disconnect explicitly in that case.

https://bugzilla.gnome.org/show_bug.cgi?id=770751