mainWindow: Fix state saving
The 'app.quit' action is implemented by calling destroy() on the window
rather than close(). As a result, the ::delete-event we currently use
for state saving won't be emitted in that case. We clearly want to
always save the state when the window is closed, not just when the
user or window manager requests that the window is closed, so use
the ::destroy event instead.