userTracker: Add global tracking
The first userTracker patch adds only the local tracking
functionality of the userTracker, but we still need the
global one too. This patch adds the global tracking
functionality. First of all we create the UserStatusMonitor,
a singleton that handles all other instances of UserTrackers.
UserTrackers are now account-specific, meaning that they
track rooms that are on the same account (network).
Obviously, each room is tracked both individually (local
tracking) and globally (global tracking). Global tracking
is achieved through emitting the status-changed::basenick
detailed signal while the local one uses custom callbacks.
The reason we use custom callbacks is that it avoids
filtering of the parameters sent by a signal. Local status
is in connection with both the basenick and the room, hence
double filtering would be required in the case of using
a detailed signal for the local tracking part.
Also, there is a new contacts-changed::basenick signal that
is emitted whenever the contacts change for a specific
basenick (useful when there are more users with the same
basenick).