userTracker: Add global tracking
This reverts commit 2d9f3605d4ef77a9dc0d8973002e2a939f6e11d9.

userTracker: *fixed rebase halfway split patch* UserTracker now extends GObject and emits detailed sinal status-changed::basenick. Add getBestMatchingContact.

userTracker: patch based on the userTracker patch that introduced the global tracking of nicks. this is more or less the *final* second patch for the userTracker.

userTracker: add the UserStatusMonitor class in order to handle all the trackers and fix conflicts in the UserTracker class.