room: Emit ::identify-sent signal when detecting authbot identification
We want to offer users that have their nickname registered with an
authbot like NickServ the option to store the password and handle
the identification automatically when connecting. For that we need
to detect the "identify" command sent to a bot first - asking the
user whether the password should be saved will be done differently
depending on whether the command is actually send from the room's
view or via a /msg command (or GNOME Shell's chat integration) instead,
so handling the detection in a shared component as the room itself
makes sense.
In the future we might want to hide authbot channels entirely from the
UI unless the user explicitly initiates the communication, in which
case the detection/signal would be better off in the channel. However
that would be more involved as it involves implementing a custom client
factor, so this should do for now ...

https://bugzilla.gnome.org/show_bug.cgi?id=709982