userTracker: add notify actions
In order to be able to show a notification whenever a
watched user comes online, we will use a GAction, as it
makes the whole process a lot cleaner. Each basenick
has an associated GAction, thus making the notifications
basenick-specific (just like all the status tracking that
is done) on that account (network). Whether or not the
action is enabled is decided by the basenick's status.
Whether or not a notification should be emitted is decided
by the state of the GAction.