telepathyClient: Ask users whether to save "identify" passwords
To identify with NickServ, users don't need to actually select the
corresponding room - they could use the /msg command or gnome-shell's
built-in chat notifications.  As the user likely won't see the info
bar added in the previous commit in that case, use a notification
to ask for confirmation when the bot's room isn't active.

https://bugzilla.gnome.org/show_bug.cgi?id=709982