telepathyClient: Cache sent identify parameters for later saving
We won't save passwords immediately when we detect an "identify" command,
but only after confirmation from the user.  In order to have the data
available for saving later, we need to keep it until the user makes a
decision - the telepathy client is a reasonable place, as it will be
available even when we don't show any UI (once we support that), and
it will be one of the components asking for confirmation.

https://bugzilla.gnome.org/show_bug.cgi?id=709982