app: Fix filtering of alternate servers
When connecting to a server of a predefined network fails, we use
the list of servers to keep trying. The intention of the code is
to skip the server that already failed, but as we currently compare
an object to a string, we end up including it anyway.
Fix the filtering function to produce the expected result.

https://bugzilla.gnome.org/show_bug.cgi?id=766935