app: Set 'use-ssl' when retrying with another server
When creating an account from a predefined network, we always use
SSL when available. Likewise, all servers return by the networks
manager will be SSL-enabled if the networks supports it. So in
theory, the 'use-ssl' parameter should already be correct and
we don't need to update it. However there have been reports
where the parameter got unset for some reason, with the result
that all connection attempts fail. While it would be interesting
to figure out how exactly the parameter is lost, setting it
anyway when retrying with another server is still a good idea,
as account could always have been modified externally.

https://bugzilla.gnome.org/show_bug.cgi?id=766935