accountsMonitor: Manage account-specific settings
Identifying with NickServ and compatible bots requires more information
than just the password, namely the username that is identified and the
name of the bot itself. Account parameters are limited to a fixed set
of settings defined by telepathy itself and the protocol, so use a
relocatable GSettings schema to store the additional parameters.

https://bugzilla.gnome.org/show_bug.cgi?id=709982