chatView: Delegate state tracking to the app
In order to only mark messages as read when we can assume that the
user has seen it, we track both whether the view's toplevel window
has focus and whether the view's room is currently active. While
the existing code to do this isn't particularly bad, it is actually
simpler to delegate the state tracking to the application, with the
added benefit that it will be available to other modules that want
to apply the same check.

https://bugzilla.gnome.org/show_bug.cgi?id=769656