application: Initialize UserStatusMonitor on startup
The UserStatusMonitor connects to the AccountsMonitor signals
in order to get notified whenever accounts are added or removed.
We want to listen to those signals right from the beginning,
because we need to able to listen for the ::room-added signal
(emitted by the RoomManager inside each UserTracker) right from
the start, as failing to receive this signal would mean that the
room is not tracked.

https://bugzilla.gnome.org/show_bug.cgi?id=760853