app: Always use a timestamp of 0 when requesting channels
Whether a join/query action should switch to the corresponding room is
determined by a timestamp contained in the action's parameter. We handle
that timestamp immediately to be responsive to user action, but currently
also pass it on to telepathy. As a result, mission-control will ask us to
present the room (again) when the channel is ready, which is pointless in
the best case. In the worst case when channel creation is slow (as during
startup, when joining multiple channels at once), the user might have
selected a different room while waiting for the connection to finish, in
which case the focus stealing is unexpected and rude.

Just use a timestamp of 0 for all channel requests we make ourselves, so
that we only present externally created channels that are handed over to
us (for example from GNOME Shell's builtin chat integration).

https://bugzilla.gnome.org/show_bug.cgi?id=769584