userPopover: notifyButton is now a toggleButton, use watchUser() in order to track status changes for the given basenick, replace getBestMatchingContact with getBestMatchingContactInRoom, use the notification logic that was moved from the chatroomManager to the userTracker.
userPopover: add the 'offline in another room' status and it is updated whenever the global status changes. Both callbacks are connected in the setter for the nickname property (this might be a bit odd, maybe find another place to put those).
userPopover: update notifyButton according to the global status of a user plus some cleaning up.

userPopover: bind the active property of the notifyButton to the isUserWatched property of the userDetails in order to show the notifyLabel whenever the notifyButton is pressed. Disconnect contactsChangedSignal. Replace getBestMatchingContactInRoom with lookupContact.

userPopover: update the active property of the notifyButton based on the new notification-emitted::basenick signal that the userTracker emits. Add _onNotificationEmitted method (will be removed in the future). We do this so that the popover can unset the toggled notifyButton.

userPopover: moved the UserPopover widgets into user-popover.ui

userPopover: bind the sensitive and visible properties of the notifyButton. Bind the active property of the notifyButton to the notifications-enabled property of the userDetails. Remove _onNotifyButtonClicked and _updateNotifyButton since the whole logic will be replaced with GActions.

userPopover: bind the notifyButton sensitivity to its visible property. Then disconnect signals when not needed anymore. Remove _onNotificationEmitted handler for the notification-emitted userTracker signal that was removed.

userPopover: nickLabel has ellipsize and max-width-chars set.

userPopover: use <small> markup for the status label and make it translatable. (will be replaced with string formats)

userPopover: Only use IDs for widgets we access

userPopover: replace watchUser() with watchRoomStatus().

userPopover: check if signal id is greater than 0, just checking its existance is not enough for consistency.

userPopover: add signal ids and disconnect them upon changing the nickname.

userPopover: global status change signal detail is now the basenick, not the nickname.

userPopover: replace nickName with baseNick.

userPopover: Comment on the popover code as well

userPopover: minor ui adjustment and only use referenced children.

userPopover: vertically center the notifyButton.

userPopover: all widgets in the UserPopover ui are marked as visible in order to reveal them by calling UserPopover.show(). userPopover: the _setBasenick() method is added in order to keep track of the basenick and disconnect connected events upon changing the basenick, but before connecting new ones. userPopover: refactor _updateContents() to _updateStatusLabel(), this fixes the 'available in another room' status that would appear when it shouldn't.

userPopover: moved the bindings of the notifyButton
to the template file.

userPopover: bind the expanded property of the UserDetails class
to its visible property, and do that in the ui file. The reason
for this binding is that we want the UserDetails to update
itself every time the popover is opened. If the expanded property
is always true, then the UserDetails will never get updated.

userPopover: set a fixed height of the UserPopover using the
height-request property and set the label of the fullnameLabel
to empty whenever it needs to be hidden. We cannod use its
visible property, as the space taken by it would be squashed
in that case

userPopover: removed the old css class that used to set the height of the popover status label.
userPopover: use the <small> markup for the status label.