roomList: Exclude placeholders from show_all()
For enabled accounts without rooms, we show an empty placeholder widget
so that we still get the header that allows managing the account. While
we correctly set the initial visibility, that code may run before any
parent calls show_all(), and we end up with disabled accounts in the
sidebar - we need to control the placeholders' visibility ourselves,
so exclude them from show_all().

https://bugzilla.gnome.org/show_bug.cgi?id=769592