chatView: add HoverFilterTag class and use it
We need a way of highlighting the nicknames as we hover
over them. We achieve this by creating a new tag that
wraps around the nickTag. The HoverFilterTag modifies the
opacity of the nickTag by first taking the color of the
nickTag, then modifying it with the hover_opacity param.
The reference to the filtered_tag is used only to get the
right color from the nickTag so that we can properly modify
it upon hovering the tag.