roomList: Use dedicated disconnected icon
We indicate both connection attempts and failures in the room list
headers, but we currently don't differentiate between the online and
(non-error) offline case. As the functionality in the latter case is
fairly limited though, it makes sense to have a clearer distinction
than just making entries insensitive, so use an appropriate icon in
the room list header.

https://bugzilla.gnome.org/show_bug.cgi?id=770751